working fake response from server, had to close buffer

+++ connector/src/main/java/co/kukurin/sensor/StoreMeasurementRequest.java
@@ -13,5 +13,5 @@ public class StoreMeasurementRequest {
 
   private String parameter;
 
-  private float averageValue;
+  private double averageValue;
 }
